refactor(clickhouse): split monolithic module into subpackage
Split observal-server/services/clickhouse.py (1293 lines) into a
Python subpackage with four focused modules:

- client.py: HTTP client, _query, _get_client, clickhouse_health,
  timestamp helpers (_now_ms, _normalize_ts), cache invalidation
- schema.py: INIT_SQL DDL for all tables, init_clickhouse,
  apply_resource_settings, TTL retention, resource tuning
- insert.py: insert_traces, insert_spans, insert_scores,
  insert_otel_logs, insert_audit_log, _insert_webhook_deliveries,
  insert_session_events
- query.py: query_traces, query_trace_by_id, query_spans,
  query_span_by_id, query_shim_spans_for_window, query_scores,
  query_recent_events, query_session_event_count,
  query_existing_for_dedup
- __init__.py: re-exports every public symbol; all existing
  callers continue to work without import changes

Submodules import the client module object so that
patch('services.clickhouse.client._query') intercepts all calls
from a single patch point.

Updated callers:
- worker.py: lazy _query import now points to client submodule
- tests: patch paths updated to services.clickhouse.client._query,
  services.clickhouse.client._get_client,
  services.clickhouse.client.settings, ch.schema._resource_overrides
