+++ observal-server/services/clickhouse/__init__.py
@@ -1,0 +1,86 @@
+# SPDX-FileCopyrightText: 2026 Hari Srinivasan <harisrini21@gmail.com>
+# SPDX-License-Identifier: AGPL-3.0-only
+"""ClickHouse subpackage.
+
+Re-exports every public symbol that was previously available from
+``services.clickhouse`` (the flat module).  All existing callers continue to
+work without import changes.
+"""
+
+from services.clickhouse.client import (
+    CLICKHOUSE_DB,
+    CLICKHOUSE_HTTP,
+    CLICKHOUSE_PASSWORD,
+    CLICKHOUSE_USER,
+    _get_client,
+    _invalidate_cache,
+    _normalize_ts,
+    _now_ms,
+    _query,
+    clickhouse_health,
+)
+from services.clickhouse.insert import (
+    _insert_webhook_deliveries,
+    insert_audit_log,
+    insert_otel_logs,
+    insert_scores,
+    insert_session_events,
+    insert_spans,
+    insert_traces,
+)
+from services.clickhouse.query import (
+    query_existing_for_dedup,
+    query_recent_events,
+    query_scores,
+    query_session_event_count,
+    query_shim_spans_for_window,
+    query_span_by_id,
+    query_spans,
+    query_trace_by_id,
+    query_traces,
+)
+from services.clickhouse.schema import (
+    DEFAULT_QUERY_SETTINGS,
+    INIT_SQL,
+    RESOURCE_SETTINGS_MAP,
+    _materialize_if_needed,
+    _resource_overrides,
+    apply_resource_settings,
+    init_clickhouse,
+)
+
+__all__ = [
+    "CLICKHOUSE_DB",
+    "CLICKHOUSE_HTTP",
+    "CLICKHOUSE_PASSWORD",
+    "CLICKHOUSE_USER",
+    "DEFAULT_QUERY_SETTINGS",
+    "INIT_SQL",
+    "RESOURCE_SETTINGS_MAP",
+    "_get_client",
+    "_insert_webhook_deliveries",
+    "_invalidate_cache",
+    "_materialize_if_needed",
+    "_normalize_ts",
+    "_now_ms",
+    "_query",
+    "_resource_overrides",
+    "apply_resource_settings",
+    "clickhouse_health",
+    "init_clickhouse",
+    "insert_audit_log",
+    "insert_otel_logs",
+    "insert_scores",
+    "insert_session_events",
+    "insert_spans",
+    "insert_traces",
+    "query_existing_for_dedup",
+    "query_recent_events",
+    "query_scores",
+    "query_session_event_count",
+    "query_shim_spans_for_window",
+    "query_span_by_id",
+    "query_spans",
+    "query_trace_by_id",
+    "query_traces",
+]
