+++ tests/test_clickhouse_phase1.py
@@ -135,7 +135,7 @@ def _mock_response(status_code=200, data=None):
 class TestInitClickhouse:
     @pytest.mark.asyncio
     async def test_calls_all_ddl(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response()
             await init_clickhouse()
             # +1 health check
@@ -152,7 +152,7 @@ class TestInitClickhouse:
 class TestInsertTraces:
     @pytest.mark.asyncio
     async def test_empty_list(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             await insert_traces([])
             mock_q.assert_not_called()
 
@@ -165,7 +165,7 @@ class TestInsertTraces:
             "start_time": "2026-01-01 00:00:00.000",
             "trace_type": "mcp",
         }
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response()
             await insert_traces([trace])
             mock_q.assert_called_once()
@@ -185,7 +185,7 @@ class TestInsertTraces:
             {"trace_id": f"t{i}", "project_id": "p1", "user_id": "u1", "start_time": "2026-01-01 00:00:00.000"}
             for i in range(3)
         ]
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response()
             await insert_traces(traces)
             data = mock_q.call_args[1].get("data", "")
@@ -197,7 +197,7 @@ class TestInsertTraces:
 
     @pytest.mark.asyncio
     async def test_raises_on_error(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.side_effect = Exception("connection refused")
             with pytest.raises(Exception, match="connection refused"):
                 await insert_traces(
@@ -215,7 +215,7 @@ class TestInsertTraces:
 class TestInsertSpans:
     @pytest.mark.asyncio
     async def test_empty_list(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             await insert_spans([])
             mock_q.assert_not_called()
 
@@ -230,7 +230,7 @@ class TestInsertSpans:
             "name": "my_tool",
             "start_time": "2026-01-01 00:00:00.000",
         }
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response()
             await insert_spans([span])
             sql = mock_q.call_args[0][0]
@@ -256,7 +256,7 @@ class TestInsertSpans:
             "relationships_used": 8,
             "tool_schema_valid": 1,
         }
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response()
             await insert_spans([span])
             data = mock_q.call_args[1].get("data", "")
@@ -270,7 +270,7 @@ class TestInsertSpans:
 class TestInsertScores:
     @pytest.mark.asyncio
     async def test_empty_list(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             await insert_scores([])
             mock_q.assert_not_called()
 
@@ -286,7 +286,7 @@ class TestInsertScores:
             "value": 0.95,
             "timestamp": "2026-01-01 00:00:00.000",
         }
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response()
             await insert_scores([score])
             sql = mock_q.call_args[0][0]
@@ -305,7 +305,7 @@ class TestInsertScores:
 class TestQueryTraces:
     @pytest.mark.asyncio
     async def test_basic_query(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response(data=[{"trace_id": "t1"}])
             result = await query_traces("proj1")
             assert len(result) == 1
@@ -319,7 +319,7 @@ class TestQueryTraces:
 
     @pytest.mark.asyncio
     async def test_with_filters(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response(data=[])
             await query_traces("p1", trace_type="mcp", mcp_id="m1", user_id="u1")
             sql = mock_q.call_args[0][0]
@@ -333,7 +333,7 @@ class TestQueryTraces:
 
     @pytest.mark.asyncio
     async def test_returns_empty_on_error(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.side_effect = Exception("timeout")
             result = await query_traces("p1")
             assert result == []
@@ -342,7 +342,7 @@ class TestQueryTraces:
 class TestQueryTraceById:
     @pytest.mark.asyncio
     async def test_found(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response(data=[{"trace_id": "t1"}])
             result = await query_trace_by_id("p1", "t1")
             assert result == {"trace_id": "t1"}
@@ -351,7 +351,7 @@ class TestQueryTraceById:
 
     @pytest.mark.asyncio
     async def test_not_found(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response(data=[])
             result = await query_trace_by_id("p1", "t999")
             assert result is None
@@ -360,7 +360,7 @@ class TestQueryTraceById:
 class TestQuerySpans:
     @pytest.mark.asyncio
     async def test_basic_query(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response(data=[{"span_id": "s1"}])
             result = await query_spans("p1", "t1")
             assert len(result) == 1
@@ -371,7 +371,7 @@ class TestQuerySpans:
 
     @pytest.mark.asyncio
     async def test_with_type_filter(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response(data=[])
             await query_spans("p1", "t1", span_type="tool_call", status="error")
             sql = mock_q.call_args[0][0]
@@ -382,7 +382,7 @@ class TestQuerySpans:
 class TestQuerySpanById:
     @pytest.mark.asyncio
     async def test_found(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response(data=[{"span_id": "s1"}])
             result = await query_span_by_id("p1", "s1")
             assert result == {"span_id": "s1"}
@@ -391,14 +391,14 @@ class TestQuerySpanById:
 class TestQueryScores:
     @pytest.mark.asyncio
     async def test_basic_query(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response(data=[{"score_id": "sc1"}])
             result = await query_scores("p1")
             assert len(result) == 1
 
     @pytest.mark.asyncio
     async def test_with_filters(self):
-        with patch("services.clickhouse._query", new_callable=AsyncMock) as mock_q:
+        with patch("services.clickhouse.client._query", new_callable=AsyncMock) as mock_q:
             mock_q.return_value = _mock_response(data=[])
             await query_scores("p1", trace_id="t1", source="eval", name="accuracy")
             sql = mock_q.call_args[0][0]
